refactor(graph): separate InternalGraph (internal 0..N-1) and DynamicExplorationGraph (external labels)
- Introduce deglib::graph::InternalGraph as abstract C++ base interface operating strictly on internal_index (0..N-1).
- Add deglib::DynamicExplorationGraph facade in deglib/graph.h for external users, transparently mapping external_label <-> internal_index.
- Update C++ core graphs (ReadOnlyGraph, MutableGraph, SizeBoundedGraph) to inherit from InternalGraph .
- Update Python bindings and deglib.graph wrapper to delegate external label translation to C++ InternalGraph.

+++ cpp/test/src/unit/graph/test_readonly_graph.cpp
@@ -311,7 +311,7 @@ TEST(ReadOnlyGraph, HasPathNoConnection) {
     deglib::graph::ReadOnlyGraph graph(mutable_graph.size(), 4, space, mutable_graph);
 
     auto path = graph.hasPath({0}, 4, 0.0f, 5);
-    EXPECT_EQ(path.size(), 0u);
+    EXPECT_EQ(path.size(), static_cast<size_t>(0));
 }
 
 TEST(ReadOnlyGraph, ExploreBasicIncludeEntry) {
